Move type, extract removePreferredOutlineDuplicates.ts
